requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-annoying==0.10.6
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-multiselectfield==0.1.12
djangorestframework==3.12.4
dotmap==1.3.23
idna==2.10
jsonschema==3.2.0
lxml==4.6.3
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.3.1
pkg-resources==0.0.0
pyrsistent==0.18.0
python-cas==1.5.0
pytz==2021.1
requests==2.25.1
simpleeval==0.9.10
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.2
urllib3==1.26.6
vdf==3.4
pandas
openpyxl